Refactor risk analysis template validation: move in a seperated file
