+++ packages/commons-test/src/riskAnalysisTestUtils.ts
@@ -83,118 +83,118 @@ export const validatedRiskAnalysisTemplate3_0_Pa: RiskAnalysisTemplateValidatedF
       {
         key: "purpose",
         value: "INSTITUTIONAL",
-        editable: true,
-        suggestedValues: ["INSTITUTIONAL", "COMMERCIAL"],
+        editable: false,
+        suggestedValues: [],
       },
       {
         key: "institutionalPurpose",
         value: "MyPurpose",
-        editable: true,
+        editable: false,
         suggestedValues: [],
       },
       {
         key: "otherPersonalDataTypes",
         value: "MyDataTypes",
-        editable: true,
+        editable: false,
         suggestedValues: [],
       },
       {
         key: "legalObligationReference",
         value: "somethingLegal",
-        editable: true,
-        suggestedValues: [],
-      },
-      {
-        key: "legalBasisPublicInterest",
-        value: "RULE_OF_LAW",
-        editable: true,
-        suggestedValues: ["RULE_OF_LAW", "PUBLIC_HEALTH"],
-      },
-      {
-        key: "ruleOfLawText",
-        value: "TheLaw",
-        editable: true,
+        editable: false,
         suggestedValues: [],
       },
       {
         key: "knowsDataQuantity",
         value: "NO",
-        editable: true,
+        editable: false,
         suggestedValues: ["YES", "NO"],
       },
       {
         key: "deliveryMethod",
         value: "ANONYMOUS",
-        editable: true,
+        editable: false,
         suggestedValues: ["ANONYMOUS", "PSEUDONYMOUS", "IDENTIFIED"],
       },
       {
         key: "policyProvided",
         value: "NO",
-        editable: true,
+        editable: false,
         suggestedValues: ["YES", "NO"],
       },
       {
         key: "confirmPricipleIntegrityAndDiscretion",
         value: "true",
-        editable: true,
+        editable: false,
         suggestedValues: ["true", "false"],
       },
       {
         key: "reasonPolicyNotProvided",
         value: "Because",
-        editable: true,
+        editable: false,
         suggestedValues: [],
       },
       {
         key: "doneDpia",
         value: "NO",
-        editable: true,
+        editable: false,
         suggestedValues: ["YES", "NO"],
       },
       {
         key: "dataDownload",
         value: "YES",
-        editable: true,
+        editable: false,
         suggestedValues: ["YES", "NO"],
       },
       {
         key: "confirmDataRetentionPeriod",
         value: "true",
-        editable: true,
+        editable: false,
         suggestedValues: ["true", "false"],
       },
       {
         key: "purposePursuit",
         value: "MERE_CORRECTNESS",
-        editable: true,
+        editable: false,
         suggestedValues: ["MERE_CORRECTNESS", "QUALITY_IMPROVEMENT"],
       },
       {
         key: "checkedExistenceMereCorrectnessInteropCatalogue",
         value: "true",
-        editable: true,
+        editable: false,
         suggestedValues: ["true", "false"],
       },
       {
         key: "usesThirdPartyData",
         value: "NO",
-        editable: true,
+        editable: false,
         suggestedValues: ["YES", "NO"],
       },
       {
         key: "declarationConfirmGDPR",
         value: "true",
-        editable: true,
+        editable: false,
         suggestedValues: ["true", "false"],
+      },
+      {
+        key: "publicInterestTaskText",
+        value: undefined,
+        editable: true,
+        suggestedValues: [],
+      },
+      {
+        key: "legalBasisPublicInterest",
+        value: "PUBLIC_INTEREST_TASK",
+        editable: false,
+        suggestedValues: [],
       },
     ],
     multiAnswers: [
-      { key: "personalDataTypes", values: ["OTHER"], editable: true },
+      { key: "personalDataTypes", values: ["OTHER"], editable: false },
       {
         key: "legalBasis",
         values: ["LEGAL_OBLIGATION", "PUBLIC_INTEREST"],
-        editable: true,
+        editable: false,
       },
     ],
   };
@@ -264,8 +264,8 @@ export const validatedRiskAnalysisTemplate2_0_Private: RiskAnalysisTemplateValid
       {
         key: "purpose",
         value: "INSTITUTIONAL",
-        editable: true,
-        suggestedValues: ["INSTITUTIONAL", "COMMERCIAL"],
+        editable: false,
+        suggestedValues: [],
       },
       {
         key: "institutionalPurpose",
@@ -375,7 +375,7 @@ export const validatedRiskAnalysisTemplate2_0_Private: RiskAnalysisTemplateValid
       {
         key: "legalBasis",
         values: ["LEGAL_OBLIGATION", "PUBLIC_INTEREST"],
-        editable: true,
+        editable: false,
       },
     ],
   };
