+++ packages/commons/src/risk-analysis/index.ts
@@ -3,4 +3,5 @@ export * from "./riskAnalysisValidationErrors.js";
 export * from "./riskAnalysisTemplateValidationErrors.js";
 export * from "./riskAnalysisTemplateErrors.js";
 export * from "./riskAnalysisValidation.js";
+export * from "./riskAnalysisTemplateValidation.js";
 export * from "./rules/index.js";
